Move test request factory under `testing.client` package
