+++ server/src/test/java/io/spine/core/EventsTest.java
@@ -27,12 +27,12 @@ import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import io.spine.base.Identifier;
 import io.spine.base.Time;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.given.EventsTestEnv;
 import io.spine.core.given.GivenEvent;
 import io.spine.server.event.EventFactory;
 import io.spine.string.Stringifiers;
 import io.spine.test.Tests;
+import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.type.TypeName;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
