+++ server/src/test/java/io/spine/core/RejectionsTest.java
@@ -28,8 +28,8 @@ import com.google.protobuf.util.Timestamps;
 import io.spine.base.Identifier;
 import io.spine.base.ThrowableMessage;
 import io.spine.base.Time;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.protobuf.AnyPacker;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
