+++ server/src/test/java/io/spine/server/CommandServiceTest.java
@@ -23,7 +23,6 @@ package io.spine.server;
 import com.google.common.collect.Sets;
 import com.google.protobuf.StringValue;
 import io.spine.base.Error;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Ack;
 import io.spine.core.Command;
 import io.spine.core.CommandId;
@@ -33,6 +32,7 @@ import io.spine.grpc.MemoizingObserver;
 import io.spine.protobuf.AnyPacker;
 import io.spine.server.model.ModelTests;
 import io.spine.server.transport.GrpcContainer;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
