+++ server/src/test/java/io/spine/server/SubscriptionServiceTest.java
@@ -26,7 +26,6 @@ import io.spine.client.Subscription;
 import io.spine.client.SubscriptionUpdate;
 import io.spine.client.Target;
 import io.spine.client.Targets;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.client.Topic;
 import io.spine.core.Response;
 import io.spine.server.Given.MemoizeStreamObserver;
@@ -36,6 +35,7 @@ import io.spine.server.model.ModelTests;
 import io.spine.server.stand.Stand;
 import io.spine.test.aggregate.Project;
 import io.spine.test.aggregate.ProjectId;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
