+++ server/src/test/java/io/spine/server/aggregate/given/AggregateRepositoryTestEnv.java
@@ -35,7 +35,6 @@ import com.google.protobuf.UInt64Value;
 import com.google.protobuf.util.Timestamps;
 import io.spine.base.Identifier;
 import io.spine.base.Time;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.CommandContext;
 import io.spine.core.CommandEnvelope;
 import io.spine.core.EventContext;
@@ -73,6 +72,7 @@ import io.spine.test.aggregate.event.AggTaskAdded;
 import io.spine.test.aggregate.rejection.AggCannotStartArchivedProject;
 import io.spine.test.aggregate.rejection.Rejections;
 import io.spine.testdata.Sample;
+import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.validate.BoolValueVBuilder;
 import io.spine.validate.StringValueVBuilder;
 import org.checkerframework.checker.nullness.qual.Nullable;
