+++ server/src/test/java/io/spine/server/aggregate/given/Given.java
@@ -22,7 +22,6 @@ package io.spine.server.aggregate.given;
 
 import com.google.protobuf.Message;
 import com.google.protobuf.Timestamp;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Command;
 import io.spine.core.UserId;
 import io.spine.core.given.GivenUserId;
@@ -38,6 +37,7 @@ import io.spine.test.aggregate.event.AggProjectPaused;
 import io.spine.test.aggregate.event.AggProjectStarted;
 import io.spine.test.aggregate.event.AggTaskAdded;
 import io.spine.testdata.Sample;
+import io.spine.testing.client.TestActorRequestFactory;
 
 import static io.spine.base.Time.getCurrentTime;
