+++ server/src/test/java/io/spine/server/command/CommandHandlerMethodTest.java
@@ -25,7 +25,6 @@ import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import io.spine.base.Identifier;
 import io.spine.base.ThrowableMessage;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.CommandContext;
 import io.spine.core.CommandEnvelope;
 import io.spine.server.aggregate.Aggregate;
@@ -49,6 +48,7 @@ import io.spine.server.model.ModelTests;
 import io.spine.test.reflect.ProjectId;
 import io.spine.test.reflect.command.RefCreateProject;
 import io.spine.test.reflect.event.RefProjectCreated;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
