+++ server/src/test/java/io/spine/server/commandbus/CommandValidatorViolationCheckTest.java
@@ -22,13 +22,13 @@ package io.spine.server.commandbus;
 
 import com.google.protobuf.Any;
 import io.spine.base.Time;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Command;
 import io.spine.core.CommandContext;
 import io.spine.core.CommandEnvelope;
 import io.spine.core.CommandId;
 import io.spine.protobuf.AnyPacker;
 import io.spine.test.command.CmdCreateProject;
+import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.validate.ConstraintViolation;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
