+++ server/src/test/java/io/spine/server/commandbus/ExecutorCommandSchedulerTest.java
@@ -23,10 +23,10 @@ package io.spine.server.commandbus;
 import com.google.protobuf.Duration;
 import com.google.protobuf.Timestamp;
 import io.spine.client.CommandFactory;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Command;
 import io.spine.core.CommandContext;
 import io.spine.core.given.GivenCommandContext;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
