+++ server/src/test/java/io/spine/server/procman/given/PmMessageDeliveryTestEnv.java
@@ -23,7 +23,6 @@ import com.google.common.collect.ImmutableSet;
 import com.google.protobuf.Message;
 import com.google.protobuf.StringValue;
 import io.spine.base.Identifier;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Command;
 import io.spine.core.Event;
 import io.spine.core.React;
@@ -45,6 +44,7 @@ import io.spine.test.procman.command.PmStartProject;
 import io.spine.test.procman.event.PmProjectCreated;
 import io.spine.test.procman.event.PmProjectStarted;
 import io.spine.test.procman.rejection.Rejections.PmCannotStartArchivedProject;
+import io.spine.testing.client.TestActorRequestFactory;
 import io.spine.validate.StringValueVBuilder;
 
 import java.util.List;
