+++ server/src/test/java/io/spine/server/stand/StandPostTest.java
@@ -23,7 +23,6 @@ import com.google.common.util.concurrent.MoreExecutors;
 import com.google.protobuf.StringValue;
 import io.netty.util.internal.ConcurrentSet;
 import io.spine.base.Identifier;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.CommandEnvelope;
 import io.spine.core.EventEnvelope;
 import io.spine.core.Version;
@@ -35,6 +34,7 @@ import io.spine.server.stand.given.Given;
 import io.spine.server.stand.given.Given.StandTestAggregate;
 import io.spine.server.storage.StorageFactory;
 import io.spine.test.projection.ProjectId;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
