+++ testutil-server/src/test/java/io/spine/server/aggregate/AggregateMessageDispatcherTest.java
@@ -24,11 +24,11 @@ import com.google.protobuf.FloatValue;
 import com.google.protobuf.Message;
 import com.google.protobuf.StringValue;
 import com.google.protobuf.UInt32Value;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.CommandEnvelope;
 import io.spine.core.EventEnvelope;
 import io.spine.server.aggregate.given.AggregateMessageDispatcherTestEnv.MessageLog;
 import io.spine.server.command.TestEventFactory;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
