+++ testutil-server/src/test/java/io/spine/server/command/CommandTestTest.java
@@ -30,7 +30,6 @@ import io.spine.core.Commands;
 import io.spine.core.TenantId;
 import io.spine.server.command.given.CommandTestTestEnv.TestCommandTest;
 import io.spine.time.testing.TimeTests;
-import io.spine.time.ZoneOffsets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -79,7 +78,6 @@ class CommandTestTest {
     static ActorRequestFactory newRequestFactory(Class<?> clazz) {
         return ActorRequestFactory.newBuilder()
                                   .setActor(newUuid())
-                                  .setZoneOffset(ZoneOffsets.utc())
                                   .setTenantId(TenantId.newBuilder()
                                                        .setValue(clazz.getSimpleName())
                                                        .build())
