+++ testutil-server/src/test/java/io/spine/server/procman/ProcessManagerDispatcherTest.java
@@ -21,12 +21,12 @@
 package io.spine.server.procman;
 
 import com.google.common.testing.NullPointerTester;
-import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Command;
 import io.spine.core.CommandEnvelope;
 import io.spine.core.EventEnvelope;
 import io.spine.core.RejectionEnvelope;
 import io.spine.server.command.TestEventFactory;
+import io.spine.testing.client.TestActorRequestFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
